+++ mods/BrewLAN/hook/lua/ui/help/unitdescription.lua
@@ -124,6 +124,7 @@ Description['ssl0323'] = "<LOC Unit_Description_BrewLAN_T3L014>Mobile tactical m
 Description['sel0323'] = "<LOC Unit_Description_BrewLAN_T3L014>Mobile tactical missile defense. Protection is limited to unit's operational area."
 Description['srl0323'] = "<LOC Unit_Description_BrewLAN_T3L014>Mobile tactical missile defense. Protection is limited to unit's operational area."
 Description['sel0324'] = "<LOC Unit_Description_BrewLAN_T3L015>Mobile anti-air missile launcher. Equipped with advanced scanning equipment."
+Description['srl0324'] = "<LOC Unit_Description_BrewLAN_T3L016>High-end mobile intelligense system. While stationary it is visible only to omni, and provides extensive radar and omni coverage."
 
 --Tech 3 buildings
 Description['srb2306'] = "<LOC Unit_Description_BrewLAN_T3B001>Heavy defensive tower. Attacks land- and sea-based units with a powerful sweeping laser. Does not engage aircraft or submerged units. Requires cumulative amounts of energy to maintain prolonged fire; will overheat if active for too long."