+++ mods/BrewLAN_Units/BrewTurrets/hook/loc/US/strings_db.lua
@@ -1,3 +1,7 @@
 srb5380_name="Deadshot"
 srb5380_desc="Accuracy Boost Node"
 Unit_Description_BrewLAN_Turrets_0001="Adjacent static weaponry spend longer aiming, resulting in higher accuracy. Multiple nodes stack."
+
+seb5380_name="Superhot"
+seb5380_desc="Damage Boost Node"
+Unit_Description_BrewLAN_Turrets_0002="Adjacent static weaponry deal increased damage. This greatly increases energy consumption per shot."