+++ mods/BrewLAN_Units/BrewTurrets/hook/lua/ui/help/unitdescription.lua
@@ -1,1 +1,2 @@
 Description['srb5380'] = "<LOC Unit_Description_BrewLAN_Turrets_0001>Adjacent static weaponry spend longer aiming, resulting in higher accuracy. Multiple nodes stack."
+Description['seb5381'] = "<LOC Unit_Description_BrewLAN_Turrets_0002>Adjacent static weaponry deal increased damage. This greatly increases energy consumption per shot."